+++ Neodomus-desarrollo-web/be/app/models/user.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, func, text
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, String, func, text
 from sqlalchemy.orm import Mapped, mapped_column
 from app.database import Base
 
@@ -11,7 +11,7 @@ class User(Base):
     last_name: Mapped[str] = mapped_column(String(100), nullable=False)
     id_tipo_documento_u: Mapped[int] = mapped_column(Integer, nullable=True)
     documento_usuario: Mapped[int] = mapped_column(Integer, unique=True, nullable=True)
-    telefono_usuario: Mapped[int] = mapped_column(Integer, nullable=True)
+    telefono_usuario: Mapped[int] = mapped_column(BigInteger, nullable=True)
     email: Mapped[str] = mapped_column(String(100), unique=True, index=True, nullable=False)
     password_hash: Mapped[str] = mapped_column(String(255), nullable=False)
     id_rol_u: Mapped[int] = mapped_column(Integer, ForeignKey("roles_usuario.id_rol"), nullable=True)
